homework14/project/taskmanager/requirements.txt
asgiref==3.5.1
Django==4.0.4
kafka-python==2.0.2
psycopg2-binary==2.9.3
sqlparse==0.4.2
